fix according to PR feedback from matthias
